Move the TargetPlatform interface to the embedder bundle
