+++ tycho-bundles/org.eclipse.tycho.embedder.shared/src/main/java/org/eclipse/tycho/DependencyResolutionException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2014 SAP SE and others.
+ * Copyright (c) 2015 SAP SE and others.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,21 +10,21 @@
  * Contributors:
  *    SAP SE - initial API and implementation
  *******************************************************************************/
-package org.eclipse.tycho.artifacts;
-
-import org.eclipse.tycho.ArtifactType;
+package org.eclipse.tycho;
 
 /**
- * Exception thrown for syntactically incorrect references to Eclipse artifacts, e.g. references
- * with an illegal version string or an unknown type.
- * 
- * @see ArtifactType
+ * Exception thrown when a dependency of a project cannot be resolved.
  */
-public class IllegalArtifactReferenceException extends Exception {
+public class DependencyResolutionException extends BuildFailureException {
 
     private static final long serialVersionUID = 1L;
 
-    public IllegalArtifactReferenceException(String message) {
+    public DependencyResolutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DependencyResolutionException(String message) {
         super(message);
     }
+
 }
