+++ tycho-bundles/org.eclipse.tycho.embedder.shared/src/main/java/org/eclipse/tycho/IllegalArtifactReferenceException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015 SAP SE and others.
+ * Copyright (c) 2014 SAP SE and others.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,23 +10,19 @@
  * Contributors:
  *    SAP SE - initial API and implementation
  *******************************************************************************/
-package org.eclipse.tycho.artifacts;
-
-import org.eclipse.tycho.core.shared.BuildFailureException;
+package org.eclipse.tycho;
 
 /**
- * Exception thrown when a dependency of a project cannot be resolved.
+ * Exception thrown for syntactically incorrect references to Eclipse artifacts, e.g. references
+ * with an illegal version string or an unknown type.
+ * 
+ * @see ArtifactType
  */
-public class DependencyResolutionException extends BuildFailureException {
+public class IllegalArtifactReferenceException extends Exception {
 
     private static final long serialVersionUID = 1L;
 
-    public DependencyResolutionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DependencyResolutionException(String message) {
+    public IllegalArtifactReferenceException(String message) {
         super(message);
     }
-
 }
