+++ tycho-bundles/org.eclipse.tycho.p2.tools.tests/src/test/java/org/eclipse/tycho/p2/tools/publisher/PublishProductToolTest.java
@@ -48,11 +48,11 @@ import java.util.Set;
 
 import org.eclipse.equinox.p2.metadata.IInstallableUnit;
 import org.eclipse.equinox.p2.query.QueryUtil;
+import org.eclipse.tycho.BuildFailureException;
+import org.eclipse.tycho.DependencyResolutionException;
 import org.eclipse.tycho.Interpolator;
 import org.eclipse.tycho.TargetEnvironment;
-import org.eclipse.tycho.artifacts.DependencyResolutionException;
 import org.eclipse.tycho.core.resolver.shared.DependencySeed;
-import org.eclipse.tycho.core.shared.BuildFailureException;
 import org.eclipse.tycho.p2.target.FinalTargetPlatformImpl;
 import org.eclipse.tycho.p2.target.P2TargetPlatform;
 import org.eclipse.tycho.p2.tools.publisher.facade.PublishProductTool;
