+++ tycho-core/src/test/java/org/eclipse/tycho/core/ee/ExecutionEnvironmentConfigurationTest.java
@@ -19,9 +19,9 @@ import java.util.List;
 
 import org.codehaus.plexus.logging.Logger;
 import org.codehaus.plexus.logging.console.ConsoleLogger;
+import org.eclipse.tycho.BuildFailureException;
 import org.eclipse.tycho.core.ee.shared.ExecutionEnvironmentConfiguration;
 import org.eclipse.tycho.core.ee.shared.SystemCapability;
-import org.eclipse.tycho.core.shared.BuildFailureException;
 import org.junit.Before;
 import org.junit.Test;
